requirements.txt
Flask==3.1.0
google-cloud-discoveryengine==0.13.5
line-bot-sdk==3.14.3
requests==2.32.3
debugpy # Required for debugging.